feat(certificate): Remove the Certificater interface in favor of a struct.
Currently there is a Certificater interface that is used to abstract a Certificate.
There are 3 identical implementations of Certificate, and some misuse between treating
a certificater as either a pointer or a struct. This PR simplifies this by consolidating
to a single implementation of Certificate, and removing the need for an interface all
together.

Signed-off-by: Sean Teeling <seanteeling@microsoft.com>